Add IP address change notification
This add a function that can be used to register
callback functions to monitor the change of an
IP address of a specified network interface.
It is implemented for Linux and Windows 64.
A new test executable was added to allow testing
of the functionallity on dynamically added dummy
network interfaces. This test excecutable must be
run with elevated rights.

Signed-off-by: Stefan Kimmer <skimmer@s2e-systems.com>